datamaxi::api

Trait ResultExt

pub trait ResultExt<T> {
    // Required method
    fn chain_err<F, EK>(self, callback: F) -> Result<T, Error>
       where F: FnOnce() -> EK,
             EK: Into<ErrorKind>;
}

Additional methods for Result, for easy interaction with this crate.

Required Methods

fn chain_err<F, EK>(self, callback: F) -> Result<T, Error>
where F: FnOnce() -> EK, EK: Into<ErrorKind>,

If the Result is an Err then chain_err evaluates the closure, which returns some type that can be converted to ErrorKind, boxes the original error to store as the cause, then returns a new error containing the original error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.