datamaxi::generated

Struct PremiumDetail

pub struct PremiumDetail {

    pub bid: String,
    pub d: i64,
    pub fg: f64,
    pub nfr: f64,
    pub pdp: f64,
    pub pdp15m: f64,
    pub pdp1h: f64,
    pub pdp24h: f64,
    pub pdp30m: f64,
    pub pdp4h: f64,
    pub pdp5m: f64,
    pub pmd: i64,
    pub sad: f64,
    pub sad2p: f64,
    pub sadf: f64,
    pub sb: String,
    pub sbd2p: f64,
    pub sc: String,
    pub se: String,
    pub sfr: f64,
    pub sfri: i64,
    pub sfrt: i64,
    pub shb: f64,
    pub sla: f64,
    pub sm: String,
    pub sms: bool,
    pub snd: i64,
    pub soi: f64,
    pub soich1h: f64,
    pub soich24h: f64,
    pub soich4h: f64,
    pub soivr: f64,
    pub sp: f64,
    pub spa: String,
    pub spdp15m: f64,
    pub spdp1h: f64,
    pub spdp24h: f64,
    pub spdp30m: f64,
    pub spdp4h: f64,
    pub spdp5m: f64,
    pub sq: String,
    pub st: i64,
    pub sv: f64,
    pub t: bool,
    pub tad2p: f64,
    pub tb: String,
    pub tbd: f64,
    pub tbd2p: f64,
    pub tbdf: f64,
    pub tc: String,
    pub te: String,
    pub tfr: f64,
    pub tfri: i64,
    pub tfrt: i64,
    pub thb: f64,
    pub tla: f64,
    pub tm: String,
    pub tms: bool,
    pub tnd: i64,
    pub toi: f64,
    pub toich1h: f64,
    pub toich24h: f64,
    pub toich4h: f64,
    pub toivr: f64,
    pub tp: f64,
    pub tpa: String,
    pub tpdp15m: f64,
    pub tpdp1h: f64,
    pub tpdp24h: f64,
    pub tpdp30m: f64,
    pub tpdp4h: f64,
    pub tpdp5m: f64,
    pub tq: String,
    pub tt: i64,
    pub tv: f64,

}

Fields
bid: Stringd: i64

specifies the date and the time in UTC milliseconds

fg: f64

funding gap which is difference between source and target fundingrate without funding interval consideration

nfr: f64

net fundingrate which takes funding interval into account

pdp: f64

specifies the price difference percentage between the source and target exchanges

pdp15m: f64

specifies the price difference percentage between the source and target exchanges 15m ago

pdp1h: f64

specifies the price difference percentage between the source and target exchanges 1h ago

pdp24h: f64

specifies the price difference percentage between the source and target exchanges 24h ago

pdp30m: f64

specifies the price difference percentage between the source and target exchanges 30m ago

pdp4h: f64

specifies the price difference percentage between the source and target exchanges 4h ago

pdp5m: f64

specifies the price difference percentage between the source and target exchanges 5m ago

pmd: i64

sepcifies premium duration

sad: f64

source ask depth within +2% base

sad2p: f64

specifies -2% volume depth from source exchange

sadf: f64

source ask depth within +2% quote

sb: String

specifies the base token of the source exchange

sbd2p: f64

specifies +2% volume depth from source exchange

sc: String

for amm source ticker, amm source chain

se: String

specifies the source exchange name

sfr: f64

source funding rate

sfri: i64

source funding rate interval, 1 stands for 1 hour

sfrt: i64

source fundingrate info timestamp in UTC millisecond

shb: f64

specifies highest bid from source exchange

sla: f64

specifies lowest bid from source exchange

sm: String

specifies the source market type

sms: bool

boolean if source exchange margin is supported, returned only for spot market

snd: i64

source next distribution time in UTC milliseconds

soi: f64

Open Interest snapshot — USD-denominated. Source and target sides.

soich1h: f64

OI % change over rolling 1h / 4h / 24h windows.

soich24h: f64soich4h: f64soivr: f64

OI / 24h USD quote volume. Crude “leverage per turnover” metric.

sp: f64

specifies the latest price of the source exchange in requested currency

spa: String

for amm source ticker, amm pool address

spdp15m: f64

specifies the price difference percentage of the source exchange in the last 15m

spdp1h: f64

specifies the price difference percentage of the source exchange in the last 1h

spdp24h: f64

specifies the price difference percentage of the source exchange in the last 24h

spdp30m: f64

specifies the price difference percentage of the source exchange in the last 30m

spdp4h: f64

specifies the price difference percentage of the source exchange in the last 4h

spdp5m: f64

specifies the price difference percentage of the source exchange in the last 5m

sq: String

specifies the quote token of the source exchange

st: i64

specifies the date and the time of source ticker in UTC milliseconds

sv: f64

specifies the trading volume of the source exchange in the last 24 hours in requested currency

t: bool

transferable, null if unknown

tad2p: f64

specifies -2% volume depth from target exchange

tb: String

specifies the base token of the target exchange

tbd: f64

target bid depth within -2% base

tbd2p: f64

specifies +2% volume depth from target exchange

tbdf: f64

target bid depth within -2% quote

tc: String

for amm target ticker, amm target chain

te: String

specifies the target exchange name

tfr: f64

target funding rate

tfri: i64

target funding rate interval, 1 stands for 1 hour

tfrt: i64

target fundingrate info timestamp in UTC millisecond

thb: f64

specifies highest bid from target exchange

tla: f64

specifies lowest bid from target exchange

tm: String

specifies the target market type

tms: bool

boolean if target exchange margin is supported, returned only for spot market

tnd: i64

target next distribution time in UTC millisconds

toi: f64toich1h: f64toich24h: f64toich4h: f64toivr: f64tp: f64

specifies the latest price of the target exchange

tpa: String

for amm target ticker, amm pool address

tpdp15m: f64

specifies the price difference percentage of the target exchange in the last 15m

tpdp1h: f64

specifies the price difference percentage of the target exchange in the last 1h

tpdp24h: f64

specifies the price difference percentage of the target exchange in the last 24h

tpdp30m: f64

specifies the price difference percentage of the target exchange in the last 30m

tpdp4h: f64

specifies the price difference percentage of the target exchange in the last 4h

tpdp5m: f64

specifies the price difference percentage of the target exchange in the last 5m

tq: String

specifies the quote token of the target exchange

tt: i64

specifies the date and the time of target ticker in UTC milliseconds

tv: f64

specifies the trading volume of the target exchange in the last 24 hours in requested currency