Trait datamaxi::api::Datamaxi

pub trait Datamaxi {
    // Required methods
    fn new(api_key: String) -> Self;
    fn new_with_base_url(api_key: String, base_url: String) -> Self;
}

A trait that defines the required methods for interacting with the Datamaxi+ API.

Required Methods

fn new(api_key: String) -> Self

Creates a new instance of the implementing type using the provided API key.

fn new_with_base_url(api_key: String, base_url: String) -> Self

Creates a new instance of the implementing type using the provided API key and base URL.

Object Safety

This trait is not object safe.